refactor: PluginForm
